fix(ci): unblock cross-platform builds — round 2
Round 1 cleared the surface symptoms; round 2 tackles the deeper
portability fan-out the cross-platform CI exposed:

- net_client_cross.yml looked for artifacts under build/<preset>/src/lib/
  net_client/, but AtlasOutputDirectory.cmake routes them into
  bin/<build_dir_name>/. Verify + upload paths now point at bin/.
- latency_histogram.cc:Record / QuantileMicros indexed std::array<>
  with int, tripping Apple Clang's -Wsign-conversion. Explicit
  static_cast<size_t> at both sites keeps the int loop counters and
  the unsigned indexing happy.
- heap.cc's POSIX path used std::aligned_alloc (C++17 / C11). Bionic
  doesn't ship aligned_alloc until API 28; the Android NDK preset
  targets API 24. Switched to ::posix_memalign — POSIX.1-2001, works
  on every UNIX target we care about, frees with the same free().
- ThreadPool used std::jthread + std::stop_token. Apple Clang's libc++
  on iOS still doesn't ship the C++20 jthread/stop_token pieces, so
  even the threading.h declaration broke iOS builds. Rebuilt
  ThreadPool::Impl on std::thread + manual join() in Shutdown(); the
  worker lambda never used the stop_token anyway. Dropped the dead
  SetThreadName(std::jthread&, ...) overload across the three
  threading translation units.
- child_process_linux.cc had no #if ATLAS_PLATFORM_LINUX guard, so
  it tried to compile on macOS / iOS / Android. Wrapped its body to
  match the rest of the *_linux.cc files.

Local Windows debug build clean; test_threading + test_latency_histogram
+ test_heap pass.

+++ src/lib/foundation/latency_histogram.cc
@@ -58,7 +58,7 @@ void LatencyHistogram::Record(std::chrono::microseconds latency) {
   uint64_t us = latency.count() < 0 ? 0 : static_cast<uint64_t>(latency.count());
   int idx = BucketIndex(us);
   std::lock_guard lock(mu_);
-  ++buckets_[idx];
+  ++buckets_[static_cast<std::size_t>(idx)];
   ++count_;
   if (us > max_micros_) max_micros_ = us;
 }
@@ -82,7 +82,7 @@ auto LatencyHistogram::QuantileMicros(double q) const -> double {
   if (target > count_) target = count_;
   uint64_t cum = 0;
   for (int i = 0; i < kBucketCount; ++i) {
-    cum += buckets_[i];
+    cum += buckets_[static_cast<std::size_t>(i)];
     if (cum >= target) {
       double lo = BucketLowerBoundMicros(i);
       double hi = BucketUpperBoundMicros(i);
